requirements.txt
pytest
requests
click
python-dotenv
